requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
pytest==7.4.3
pytest-cov==4.1.0